+++ test/envtest/gateway-api/gateway_envtest_test.go
@@ -173,6 +173,7 @@ func TestGatewayInfrastructureLabels(t *testing.T) {
 			Scheme:                scheme,
 			Namespace:             ns.Name,
 			DefaultDataPlaneImage: consts.DefaultDataPlaneImage,
+			ReferenceGrantVersion: envtest.ReferenceGrantVersion,
 		},
 		&dpreconciler.Reconciler{
 			DataPlaneIPFamily:        ipfamily.IPv4,