+++ test/envtest/gateway-api/gateway_grpcroute_envtest_test.go
@@ -39,6 +39,7 @@ func TestGatewayGRPCRouteAttachedRoutes(t *testing.T) {
 		Scheme:                scheme,
 		Namespace:             ns.Name,
 		DefaultDataPlaneImage: "kong:latest",
+		ReferenceGrantVersion: envtest.ReferenceGrantVersion,
 	}
 	envtest.StartReconcilers(ctx, t, mgr, logs, r)