+++ test/envtest/gateway_udproute_envtest_test.go
@@ -38,6 +38,7 @@ func TestGatewayUDPRouteAttachedRoutes(t *testing.T) {
 		Scheme:                scheme,
 		Namespace:             ns.Name,
 		DefaultDataPlaneImage: "kong:latest",
+		ReferenceGrantVersion: ReferenceGrantVersion,
 	}
 	StartReconcilers(ctx, t, mgr, logs, r)